requirements.txt
python-telegram-bot>=20.0
anthropic>=0.40.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
